updated model and added javadoc
